requirements.txt
asgiref==3.5.0
Django==3.2
django-filter==21.1
django-rest-framework-nested==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.10.1
Markdown==3.3.6
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
